+++ example/params.ts
@@ -2,7 +2,8 @@ import KingWorld from '../src'
 
 const app = new KingWorld()
 	.get('/', () => 'KINGWORLD')
-	.get('/id/:id', ({ params }) => params.id || 'No Params')
+	// Retrieve params, automatically typed
+	.get('/id/:id', ({ params }) => params.id)
 	.listen(8080)
 
 console.log('Listen')