+++ example/pure.ts
@@ -1,3 +1,4 @@
+// Using navitve Bun http server
 Bun.serve({
 	port: 8080,
 	fetch: (request: Request) => new Response('Hi'),