+++ src/index.ts
@@ -7,7 +7,15 @@ import { zodToJsonSchema } from 'zod-to-json-schema'
 import { Router } from './router'
 import Context from './context'
 import { mapResponse, mapEarlyResponse } from './handler'
-import { mapQuery, getPath, clone, mergeHook, mergeDeep, SCHEMA } from './utils'
+import {
+	mapQuery,
+	getPath,
+	clone,
+	mergeHook,
+	mergeDeep,
+	SCHEMA,
+	formatAjvError
+} from './utils'
 import { registerSchemaPath } from './schema'
 import KingWorldError from './error'
 
@@ -32,12 +40,16 @@ import {
 	LifeCycleEvent,
 	LifeCycleStore,
 	VoidLifeCycle,
-	AfterRequestHandler
+	AfterRequestHandler,
+	SchemaValidator
 } from './types'
 
 export default class KingWorld<
-	Instance extends KingWorldInstance = KingWorldInstance
+	Instance extends KingWorldInstance = KingWorldInstance,
+	InheritedSchema extends TypedSchema = TypedSchema
 > {
+	private config: KingWorldConfig
+
 	store: Instance['store'] = {
 		[SCHEMA]: {}
 	}
@@ -65,8 +77,8 @@ export default class KingWorld<
 	}
 
 	server: Server | null = null
+	private $schema: SchemaValidator | null = null
 
-	private config: KingWorldConfig
 	private router = new Router()
 	protected routes: InternalRoute<Instance>[] = []
 
@@ -179,7 +191,7 @@ export default class KingWorld<
 		return this
 	}
 
-	afterHandle<Route extends TypedRoute = TypedRoute>(
+	onAfterHandle<Route extends TypedRoute = TypedRoute>(
 		handler: AfterRequestHandler<Route, Instance>
 	) {
 		this.event.afterHandle.push(handler)
@@ -262,16 +274,21 @@ export default class KingWorld<
 		return this
 	}
 
-	guard(
-		hook: RegisterHook<{}, Instance>,
-		run: (group: KingWorld<Instance>) => void
+	guard<Schema extends TypedSchema = InheritedSchema>(
+		hook: LocalHook<Schema, Instance>,
+		run: (group: KingWorld<Instance, Schema>) => void
 	) {
-		const instance = new KingWorld<Instance>()
+		const instance = new KingWorld<Instance, Schema>()
 		run(instance)
 
 		Object.values(instance.routes).forEach(
-			({ method, path, handler, hooks: localHooks }) => {
-				this._addHandler(method, path, handler, localHooks as any)
+			({ method, path, handler, hooks: localHook }) => {
+				this._addHandler(
+					method,
+					path,
+					handler,
+					mergeHook(hook as LocalHook<any>, localHook)
+				)
 			}
 		)
 
@@ -290,7 +307,10 @@ export default class KingWorld<
 		return plugin(this as unknown as any, config) as unknown as any
 	}
 
-	get<Schema extends TypedSchema = TypedSchema, Path extends string = string>(
+	get<
+		Schema extends TypedSchema = InheritedSchema,
+		Path extends string = string
+	>(
 		path: Path,
 		handler: LocalHandler<Schema, Instance, Path>,
 		hook?: LocalHook<Schema, Instance>
@@ -301,7 +321,7 @@ export default class KingWorld<
 	}
 
 	post<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -313,7 +333,10 @@ export default class KingWorld<
 		return this
 	}
 
-	put<Schema extends TypedSchema = TypedSchema, Path extends string = string>(
+	put<
+		Schema extends TypedSchema = InheritedSchema,
+		Path extends string = string
+	>(
 		path: Path,
 		handler: LocalHandler<Schema, Instance, Path>,
 		hook?: LocalHook<Schema, Instance>
@@ -324,7 +347,7 @@ export default class KingWorld<
 	}
 
 	patch<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -337,7 +360,7 @@ export default class KingWorld<
 	}
 
 	delete<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -350,7 +373,7 @@ export default class KingWorld<
 	}
 
 	options<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -363,7 +386,7 @@ export default class KingWorld<
 	}
 
 	head<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -376,7 +399,7 @@ export default class KingWorld<
 	}
 
 	trace<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -389,7 +412,7 @@ export default class KingWorld<
 	}
 
 	connect<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		path: Path,
@@ -402,7 +425,7 @@ export default class KingWorld<
 	}
 
 	method<
-		Schema extends TypedSchema = TypedSchema,
+		Schema extends TypedSchema = InheritedSchema,
 		Path extends string = string
 	>(
 		method: HTTPMethod,
@@ -423,12 +446,15 @@ export default class KingWorld<
 				? AsyncReturned
 				: Returned
 			: Value,
-		NewInstance = KingWorld<{
-			store: Instance['store'] & {
-				[key in Key]: ReturnValue
-			}
-			request: Instance['request']
-		}>
+		NewInstance = KingWorld<
+			{
+				store: Instance['store'] & {
+					[key in Key]: ReturnValue
+				}
+				request: Instance['request']
+			},
+			InheritedSchema
+		>
 	>(name: Key, value: Value): NewInstance {
 		;(this.store as Record<Key, Value>)[name] = value
 
@@ -438,14 +464,32 @@ export default class KingWorld<
 	decorate<
 		Name extends string,
 		Callback extends Function = () => unknown,
-		NewInstance = KingWorld<{
-			store: Instance['store']
-			request: Instance['request'] & { [key in Name]: Callback }
-		}>
+		NewInstance = KingWorld<
+			{
+				store: Instance['store']
+				request: Instance['request'] & { [key in Name]: Callback }
+			},
+			InheritedSchema
+		>
 	>(name: Name, value: Callback): NewInstance {
 		return this.onTransform(
 			(app: any) => (app[name] = value)
 		) as unknown as NewInstance
+	}
+
+	schema<
+		Schema extends TypedSchema = InheritedSchema,
+		NewInstance = KingWorld<Instance, Schema>
+	>(schema: Schema): NewInstance {
+		this.$schema = {
+			body: this.getSchemaValidator(schema?.body),
+			header: this.getSchemaValidator(schema?.header),
+			params: this.getSchemaValidator(schema?.params),
+			query: this.getSchemaValidator(schema?.query),
+			response: this.getSchemaValidator(schema?.response)
+		}
+
+		return this as unknown as NewInstance
 	}
 
 	// ? Using arrow to bind `this`
@@ -495,9 +539,8 @@ export default class KingWorld<
 			if (response instanceof Promise) response = await response
 		}
 
-		if (handler.validator) {
-			const validate = handler.validator
-
+		const validate = handler.validator ?? this.$schema
+		if (validate) {
 			if (validate.header) {
 				const _header: Record<string, string> = {}
 
@@ -506,25 +549,29 @@ export default class KingWorld<
 
 				validate.header(_header)
 
-				if (validate.header.errors) throw validate.header.errors[0]
+				if (validate.header.errors)
+					throw formatAjvError('header', validate.header.errors[0])
 			}
 
 			if (validate.params) {
 				validate.params(context.params)
 
-				if (validate.params.errors) throw validate.params.errors[0]
+				if (validate.params.errors)
+					throw formatAjvError('params', validate.params.errors[0])
 			}
 
 			if (validate.query) {
 				validate.query(context.query)
 
-				if (validate.query.errors) throw validate.query.errors[0]
+				if (validate.query.errors)
+					throw formatAjvError('query', validate.query.errors[0])
 			}
 
 			if (validate.body) {
 				validate.body(body)
 
-				if (validate.body.errors) throw validate.body.errors[0]
+				if (validate.body.errors)
+					throw formatAjvError('body', validate.body.errors[0])
 			}
 		}
 
@@ -546,11 +593,10 @@ export default class KingWorld<
 		let response = handler.handle(context)
 		if (response instanceof Promise) response = await response
 
-		if (handler.validator?.response) {
-			handler.validator.response(response)
+		if (validate?.response) {
+			validate.response(response)
 
-			if (handler.validator.response.errors)
-				throw handler.validator.response.errors[0]
+			if (validate.response.errors) throw validate.response.errors[0]
 		}
 
 		for (let i = 0; i < handler.hooks.afterHandle.length; i++) {
@@ -565,9 +611,12 @@ export default class KingWorld<
 	}
 
 	private handleError = (err: Error) => {
-		const error = new KingWorldError(err.message as unknown as ErrorCode, {
-			cause: err.cause
-		})
+		const error =
+			err instanceof KingWorldError
+				? err
+				: new KingWorldError(err.name as any, err.message, {
+						cause: err.cause
+				  })
 
 		for (let i = 0; i < this.event.error.length; i++) {
 			const response = this.event.error[i](error)
@@ -590,6 +639,11 @@ export default class KingWorld<
 					status: 404
 				})
 
+			case 'VALIDATION':
+				return new Response(error.message, {
+					status: 400
+				})
+
 			default:
 				return new Response(error.message, {
 					status: 500
@@ -597,10 +651,11 @@ export default class KingWorld<
 		}
 	}
 
-	listen(options: number | Omit<Serve, 'fetch'>) {
+	listen = (options: number | Omit<Serve, 'fetch'>) => {
 		if (!Bun) throw new Error('Bun to run')
 
-		for (let i = 0; i < this.event.start.length; i++) this.event.start[i]()
+		for (let i = 0; i < this.event.start.length; i++)
+			this.event.start[i](this)
 
 		this.server = Bun.serve(
 			typeof options === 'number'
@@ -625,12 +680,12 @@ export default class KingWorld<
 				"KingWorld isn't running. Call `app.listen` to start the server."
 			)
 
-		this.server.stop()
-
 		for (let i = 0; i < this.event.stop.length; i++) {
-			const process = this.event.stop[i]()
+			const process = this.event.stop[i](this)
 			if (process instanceof Promise) await process
 		}
+
+		this.server.stop()
 	}
 }